Improve (CAPA) upgrade tests reliability
